Move vscode helpers to public/node
