+++ packages/cli-kit/src/index.ts
@@ -18,7 +18,6 @@ export * as system from './system.js'
 export * as template from './template.js'
 export * as toml from './toml.js'
 export * as ui from './ui.js'
-export * as vscode from './vscode.js'
 export * as outputMocker from './testing/output.js'
 export * as metadata from './metadata.js'
 export * as monorail from './monorail.js'
